+++ packages/core/streams/stream-data.ts
@@ -19,6 +19,8 @@ export class experimental_StreamData {
 
   // array to store appended data
   private data: JSONValue[] = [];
+  private messageAnnotations: JSONValue[] = [];
+
   constructor() {
     this.isClosedPromise = new Promise(resolve => {
       this.isClosedPromiseResolver = resolve;
@@ -37,6 +39,14 @@ export class experimental_StreamData {
           );
           self.data = [];
           controller.enqueue(encodedData);
+        }
+
+        if (self.messageAnnotations.length) {
+          const encodedMessageAnnotations = self.encoder.encode(
+            formatStreamPart('message_annotations', self.messageAnnotations),
+          );
+          self.messageAnnotations = [];
+          controller.enqueue(encodedMessageAnnotations);
         }
 
         controller.enqueue(chunk);
@@ -61,6 +71,13 @@ export class experimental_StreamData {
         if (self.data.length) {
           const encodedData = self.encoder.encode(
             formatStreamPart('data', self.data),
+          );
+          controller.enqueue(encodedData);
+        }
+
+        if (self.messageAnnotations.length) {
+          const encodedData = self.encoder.encode(
+            formatStreamPart('message_annotations', self.messageAnnotations),
           );
           controller.enqueue(encodedData);
         }
@@ -94,9 +111,15 @@ export class experimental_StreamData {
       throw new Error('Data Stream has already been closed.');
     }
 
-    this.controller?.enqueue(
-      this.encoder.encode(formatStreamPart('audio', speechBase64)),
-    );
+    this.encoder.encode(formatStreamPart('audio', speechBase64));
+  }
+
+  appendMessageAnnotation(value: JSONValue): void {
+    if (this.isClosed) {
+      throw new Error('Data Stream has already been closed.');
+    }
+
+    this.messageAnnotations.push(value);
   }
 }